hashing done

+++ backend/controllers/adminController.js
@@ -1,14 +1,17 @@
 const asyncHandler = require("express-async-handler")
 const Advice = require('../models/adviceModel')
 const Stories = require('../models/storiesModel')
-const { request } = require('../config/dbConfig')
+const { pool,sql } = require('../config/dbConfig')
 
-//POST
+//PATCH
 const approveProfile = asyncHandler(async (req, res) => {
 
     const id = req.params.id
 
     try {
+        
+        const request = pool.request()
+
         await request
             .input('id', sql.Int, id)
             .execute('ApproveProfile')
@@ -27,23 +30,22 @@ const deleteAlumnusProfile = asyncHandler(async (req, res) => {
 
     const id = req.params.id
 
-    //console.log(config)
-
     try {
-        let result = await request
-            .query(`SELECT * FROM alumni_profile WHERE id = ${id}`)
+        
+        const request = pool.request()
 
-        if (result.recordset[0] == null) {
-            res.status(400).json({ message: `Alumnus with id = ${id} does not exist` })
-        }
-
-        result = await request
+        const result = await request
             .input('id', sql.Int, id)
             .execute('DeleteAlumnus')
 
-        Advice.deleteOne({ ERP: id })
-        Stories.deleteOne({ ERP: id })
-        res.status(200).json({ message: "Alumnus successfully eradicated" })
+        if (result.returnValue === 1){
+            Advice.deleteOne({ ERP: id })
+            Stories.deleteOne({ ERP: id })
+            res.status(200).json({ message: "Alumnus successfully eradicated" })
+        }
+        else {
+            res.status(400).json({ message: `Profile with ID = ${id} does not exist` })
+        }
 
     }
     catch (err) {
@@ -58,19 +60,21 @@ const deleteStudentProfile = asyncHandler(async (req, res) => {
     const id = req.params.id
 
     try {
-        let result = await request
-            .query(`SELECT * FROM student_profile WHERE id = ${id}`)
+        
+        const request = pool.request()
 
-        if (result.recordset[0] == null) {
-            res.status(400).json({ message: `Student with id = ${id} does not exist` })
-        }
-
-        await request
+        const result = await request
             .input('id', sql.Int, id)
             .execute('DeleteStudent')
 
-        res.status(200).json({ message: "Student successfully eradicated" })
+        if (result.returnValue === 1){
+            res.status(200).json({ message: "Student successfully eradicated" })
+        }
+        else {
+            res.status(400).json({ message: `Profile with ID = ${id} does not exist` })
+        }
 
+        
     }
     catch (err) {
         console.log(`Error executing query: ${err}`)
